requirements.txt
aiosmtpd==1.4.6
alabaster==1.0.0
alembic==1.18.4
atpublic==7.0.0
attrs==26.1.0
babel==2.18.0
bcrypt==5.0.0
bidict==0.23.1
blinker==1.9.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.2
colorama==0.4.6
Deprecated==1.3.1
dnspython==2.8.0
docutils==0.22.4
email-validator==2.3.0
Flask==3.1.3
Flask-Bcrypt==1.0.1
Flask-HTTPAuth==4.8.1
Flask-Limiter==4.1.1
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-Migrate==4.1.0
Flask-Moment==1.0.6
Flask-SocketIO==5.6.1
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
gevent==26.4.0
gevent-websocket==0.10.1
greenlet==3.4.0
gunicorn==25.3.0
h11==0.16.0
idna==3.11
imagesize==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
limits==5.8.0
Mako==1.3.10
MarkupSafe==3.0.3
ordered-set==4.1.0
packaging==26.0
psycopg2-binary==2.9.11
pycparser==3.0
Pygments==2.20.0
PyJWT==2.12.1
python-dotenv==1.2.2
python-engineio==4.13.1
python-socketio==5.16.1
redis==7.4.0
requests==2.33.1
roman-numerals==4.1.0
simple-websocket==1.1.0
snowballstemmer==3.0.1
Sphinx==9.1.0
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
SQLAlchemy==2.0.49
typing_extensions==4.15.0
urllib3==2.6.3
Werkzeug==3.1.8
wrapt==2.1.2
wsproto==1.3.2
WTForms==3.2.1
zope.event==6.1
zope.interface==8.3
